adding details screen WIP
1-adding details screen
2-adding details screen viewModel
3-adding details screen needed useCases
4-adding details screen views

+++ wishList/src/main/java/com/example/wishlist/navigation/WishListNavigation.kt
@@ -10,7 +10,7 @@ const val WISH_LIST_ROUTE = "WishList"
 
 fun NavGraphBuilder.wishList(
     navigateToHomeScreen:()->Unit,
-    navigateToDetailsScreen:()->Unit
+    navigateToDetailsScreen:(Int)->Unit
 ) {
     composable(WISH_LIST_ROUTE) {
         val viewModel:WishListViewModel= hiltViewModel()
